backend/requirements.txt
fastapi
sqlalchemy
asyncpg
psycopg2-binary
typer
uvicorn
python-multipart